realtime-dashboard/requirements.txt
dash==2.17.1
plotly==5.17.0
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2
dash-bootstrap-components==1.5.0
yfinance==0.2.18
python-dotenv==1.0.0
psutil==5.9.8
